requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
click==8.3.2
colorama==0.4.6
fastapi==0.136.0
filelock==3.29.0
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
packaging==26.1
pillow==12.2.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
python-multipart==0.0.26
PyYAML==6.0.3
regex==2026.4.4
rich==15.0.0
safetensors==0.7.0
setuptools==81.0.0
shellingham==1.5.4
starlette==1.0.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.6.2
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.45.0
